Correções em coordinator e Routes, ajustes de estilização
