+++ modulo3/projeto-labex/src/components/CardListTrips/CardListTrips-styled.js
@@ -1,0 +1,17 @@
+import styled from 'styled-components'
+
+export const CardTrips = styled.div`
+border-radius: 8px;
+width: 50%;
+min-height: 140px;
+margin: 4px 0;
+padding: 8px 20px;
+background-color: white;
+display: flex;
+flex-direction: column;
+justify-content: space-evenly;
+h3 {
+    margin-bottom: 8px;
+    color: #3333AA;
+}
+`
