+++ modulo3/projeto-labex/src/constants/GlobalStyles.js
@@ -1,0 +1,13 @@
+import { createGlobalStyle } from 'styled-components'
+
+export const GlobalStyles = createGlobalStyle`
+* {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+  
+}
+p {
+  color: #48404F;
+}
+`
