+++ modulo3/projeto-labex/src/pages/ApplicationFormPage/ApplicationFormPage.js
@@ -2,6 +2,6 @@ import React from 'react'
 
 export default function ApplicationFormPage() {
     return <div>
-        <p>Formulário</p>
+        Formulário
     </div>
 }
