+++ modulo3/projeto-labex/src/routes/Router.js
@@ -1,22 +1,23 @@
 import { BrowserRouter, Routes, Route } from 'react-router-dom'
-import HomePage from "../pages/HomePage"
-import ListTripsPage from "../pages/ListTripsPage"
-import AdminHomePage from '../pages/AdminHomePage'
-import ApplicationFormPage from '../pages/ApplicationFormPage'
-import CreateTripPage from '../pages/CreateTripPage'
-import LoginPage from '../pages/LoginPage'
-import TripDetailsPage from '../pages/TripDetailsPage'
+import HomePage from "../pages/HomePage/HomePage"
+import ListTripsPage from "../pages/ListTripsPage/ListTripsPage"
+import AdminHomePage from '../pages/AdminHomePage/AdminHomePage'
+import ApplicationFormPage from '../pages/ApplicationFormPage/ApplicationFormPage'
+import CreateTripPage from '../pages/CreateTripPage/CreateTripPage'
+import LoginPage from '../pages/LoginPage/LoginPage'
+import TripDetailsPage from '../pages/TripDetailsPage/TripDetailsPage'
 
 
 export const Router = () => {
     return <BrowserRouter>
         <Routes>
             <Route index element={<HomePage />} />
-            <Route path="viagens" element={<ListTripsPage />} />
-            <Route path="admin" element={<AdminHomePage />} />
-            <Route path="inscrever" element={<ApplicationFormPage />} />
-            <Route path="login" element={<LoginPage />} />
-            {/* <Route path="" element={} /> */}
+            <Route path="/trips/list" element={<ListTripsPage />} />
+            <Route path="/trips/application" element={<ApplicationFormPage />} />
+            <Route path="/login" element={<LoginPage />} />
+            <Route path="/admin/trips/list" element={<AdminHomePage />} />
+            <Route path="/admin/trips/create" element={<CreateTripPage />} />
+            <Route path="/admin/trips/:id" element={<TripDetailsPage />} />
         </Routes>
     </BrowserRouter>
 }
