+++ modulo3/projeto-labex/src/routes/coordinator.js
@@ -5,20 +5,20 @@ export const goBack = (navigate) => {
     navigate(-1)
 }
 export const goToListTripsPage = (navigate) => {
-    navigate("/viagens")
-}
-export const goToAdminHomePage = (navigate) => {
-    navigate("/admin")
+    navigate("/trips/list")
 }
 export const goToApplicationFormPage = (navigate) => {
-    navigate("/inscrever")
-}
-export const goToCreateTripPage= (navigate) => {
-    navigate("/criar")
+    navigate("/trips/application")
 }
 export const goToLoginPage= (navigate) => {
     navigate("/login")
 }
+export const goToAdminHomePage = (navigate) => {
+    navigate("/admin/trips/list")
+}
+export const goToCreateTripPage= (navigate) => {
+    navigate("/admin/trips/create")
+}
 export const goToTripDetailsPage= (navigate, id) => {
-    navigate(`/detalhes/${id}`)
+    navigate(`/admin/trips/${id}`)
 }
